fix: Resolve PHP linting errors in test files
- Fix whitespace at end of line in docblock
- Move MockErrorHandler to separate file to comply with PSR-1
- Add proper import for MockErrorHandler class

All tests continue to pass and linting errors are resolved.
